Dont send notification to originator, use Watchers
Now we dont send notifications to whoever originated the event, as those
are superfluous.

We also use the Watchers instead of the Reviewers of the commit; every
reviewer is automatically added to the Watchers list, as is the creator
of the merge request. This looks like the proper way to decide who gets
notified.